Rewrite of the footer
Added TLDR of myself, and 3 columns with navigation, donations and
contact sections.

The TLDR has two different styles depending on screen size, by default
is going to show the logo on top of the text, but for bigger screens
it's going to show up behind the text and with some opacity applied to
it.

I also did some changes to fonts in general and small updates to global
styles.